Clean up code and organize logic
