--- src/app/alertService.js
@@ -1,4 +1,0 @@
-
-const sayName = (name) => `Holla! My name is ${name}`;
-
-export default sayName;
