--- src/app/componentService.js
@@ -1,7 +1,0 @@
-import toDoItem from './app.js';
-
-const loadTodo = (()=>{
-
-})()
-
-// export { component, render };
