+++ src/index.js
@@ -1,10 +1,37 @@
 import './main.scss';
-import defaultIndex from './app/defaultIndex'
-// import toDoItem from './app/app';
-// import './app/alertService';
-// import { render } from './app/componentService';
+import contentCreator from './helpers/contentCreator';
+import addTodos from "./app/addList";
+import addProject from "./app/addProject";
+import { clearContent, createDefaultProject, displayProjectNames, getToDoTitles } from './helpers/library'
 
+console.log(localStorage)
 
+const content = document.getElementById('content')
 
+const leftSide = contentCreator.withText('div', '', 'left-side')
+content.appendChild(leftSide)
 
-defaultIndex(document.getElementById('content'))
+leftSide.appendChild(contentCreator.withText('h2', 'Project List'));
+
+createDefaultProject()
+
+displayProjectNames(leftSide)
+
+const rightSide = contentCreator.withText('div', '', 'right-side')
+content.appendChild(rightSide)
+
+const addNewList = contentCreator.withText('button', "Add New List", 'addNewList')
+addNewList.onclick = () => {
+  clearContent(content)
+  content.appendChild(addTodos())
+}
+rightSide.appendChild(addNewList)
+
+const addNewProject = contentCreator.withText('button', "Add New Project", 'addNewProject')
+addNewProject.onclick = () => {
+  clearContent(content)
+  content.appendChild(addProject())
+}
+rightSide.appendChild(addNewProject)
+
+content.classList.add('default');
